casi hecho

+++ libft_l/ft_strncpy.c
@@ -3,18 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strncpy.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: alluengo <alluengo@student.42madrid.c      +#+  +:+       +#+        */
+/*   By: alluengo <alluengo@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/20 17:30:40 by alluengo          #+#    #+#             */
-/*   Updated: 2025/08/21 20:16:03 by alluengo         ###   ########.fr       */
+/*   Updated: 2025/10/07 11:36:45 by alluengo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <unistd.h>
+#include "libft.h"
 
-char	*ft_strncpy(char *dest, char *src, unsigned int n)
+char	*ft_strncpy(char *dest, char *src, size_t n)
 {
-	unsigned int	i;
+	unsigned long	i;
 
 	i = 0;
 	while (src[i] && (i < n))
